Aufgabe 3

+++ include/DiGraph.h
@@ -17,7 +17,7 @@ public:
 
     void addNode(Node * node);
 
-    void addEdge(std::string pKey1, std::string pKey2, float pWeight);
+    void addEdge(std::string pKey1, std::string pKey2, double pWeight);
 
     Liste<Node*> getNeighbours(std::string key);
